moved roles to const utils
